[Feature][Engine] Add spark engine error data sink
