+++ datavines-engine/datavines-engine-plugins/datavines-engine-jdbc/datavines-engine-jdbc-config/src/main/java/io/datavines/engine/jdbc/config/BaseJdbcConfigurationBuilder.java
@@ -116,57 +116,5 @@ public abstract class BaseJdbcConfigurationBuilder extends BaseDataQualityConfig
         return sourceConfigs;
     }
 
-    @Override
-    public void buildTransformConfigs() {
 
-        String metricType = taskParameter.getMetricType();
-        SqlMetric sqlMetric = PluginLoader
-                .getPluginLoader(SqlMetric.class)
-                .getNewPlugin(metricType);
-
-        MetricParserUtils.operateInputParameter(inputParameter, sqlMetric, taskInfo);
-
-        List<TransformConfig> transformConfigs = new ArrayList<>();
-
-        inputParameter.put(INVALIDATE_ITEMS_TABLE,
-                sqlMetric.getInvalidateItems().getResultTable()
-                        + "_" + inputParameter.get(ConfigConstants.TASK_ID));
-
-        MetricParserUtils.setTransformerConfig(
-                inputParameter, transformConfigs,
-                sqlMetric.getInvalidateItems(), TransformType.INVALIDATE_ITEMS.getDescription());
-
-        MetricParserUtils.setTransformerConfig(
-                inputParameter,
-                transformConfigs,
-                sqlMetric.getActualValue(),
-                TransformType.ACTUAL_VALUE.getDescription());
-
-        inputParameter.put(ACTUAL_TABLE, sqlMetric.getActualValue().getResultTable());
-
-        // get expected value transform sql
-        String expectedType = taskInfo.getEngineType() + "_" + taskParameter.getExpectedType();
-        expectedValue = PluginLoader
-                .getPluginLoader(ExpectedValue.class)
-                .getNewPlugin(expectedType);
-
-        ExecuteSql expectedValueExecuteSql =
-                new ExecuteSql(expectedValue.getExecuteSql(),expectedValue.getOutputTable());
-
-        if (StringUtils.isNotEmpty(expectedValueExecuteSql.getResultTable())) {
-            inputParameter.put(EXPECTED_TABLE, expectedValueExecuteSql.getResultTable());
-        }
-
-        inputParameter.put(UNIQUE_CODE, StringUtils.wrapperSingleQuotes(generateUniqueCode(inputParameter)));
-
-        if (expectedValue.isNeedDefaultDatasource()) {
-            MetricParserUtils.setTransformerConfig(inputParameter, transformConfigs,
-                    expectedValueExecuteSql, TransformType.EXPECTED_VALUE_FROM_METADATA_SOURCE.getDescription());
-        } else {
-            MetricParserUtils.setTransformerConfig(inputParameter, transformConfigs,
-                    expectedValueExecuteSql, TransformType.EXPECTED_VALUE_FROM_SOURCE.getDescription());
-        }
-
-        configuration.setTransformParameters(transformConfigs);
-    }
 }
