+++ datavines-engine/datavines-engine-plugins/datavines-engine-spark/datavines-engine-spark-config/src/main/java/io/datavines/engine/spark/config/SparkMultiTableAccuracyMetricBuilder.java
@@ -61,6 +61,8 @@ public class SparkMultiTableAccuracyMetricBuilder extends BaseSparkConfiguration
         //todo
         //get the error data storage parameter
         //support file(hdfs/minio/s3)/es
+        SinkConfig errorDataSinkConfig = getErrorSinkConfig();
+        sinkConfigs.add(errorDataSinkConfig);
 
         configuration.setSinkParameters(sinkConfigs);
     }
