+++ datavines-engine/datavines-engine-plugins/datavines-engine-spark/datavines-engine-spark-config/src/main/java/io/datavines/engine/spark/config/SparkSingleTableCustomSqlMetricBuilder.java
@@ -74,6 +74,8 @@ public class SparkSingleTableCustomSqlMetricBuilder extends BaseSparkConfigurati
         //todo
         //get the error data storage parameter
         //support file(hdfs/minio/s3)/es
+        SinkConfig errorDataSinkConfig = getErrorSinkConfig();
+        sinkConfigs.add(errorDataSinkConfig);
 
         configuration.setSinkParameters(sinkConfigs);
     }
