+++ datavines-engine/datavines-engine-plugins/datavines-engine-spark/datavines-engine-spark-config/src/main/java/io/datavines/engine/spark/config/SparkSingleTableMetricBuilder.java
@@ -49,6 +49,8 @@ public class SparkSingleTableMetricBuilder extends BaseSparkConfigurationBuilder
         //todo
         //get the error data storage parameter
         //support file(hdfs/minio/s3)/es
+        SinkConfig errorDataSinkConfig = getErrorSinkConfig();
+        sinkConfigs.add(errorDataSinkConfig);
 
         configuration.setSinkParameters(sinkConfigs);
     }
